Create folders, index.html, styles.css
The fundamental folders of every web project is created
(media, pages, styles) folders to keep development organized in the long
term

index.html are divided into headers, sections, and footers with
classes corressponding to its part in the design image for increased
visibility & making CSS easier to code. index.html will use style.css in
the "styles/" folder

styles.css includes a CSS reset to get rid of the browser default
padding, margin, and box-sizing. This is to gain more control over the
style we want to implement