Add color & fonts CSS and replace classes in HTML
The CSS now contain all of color & font styles from the design.
All color & fonts are grouped into classes named after the element for
its intended uses. It's commented in the CSS as a "Style Classes".

Whenever we want to apply a style on an element, we can call upon that
"Style Classes" in the HTML element to apply that style - making it a
clear distinction between which CSS affect the "layout" & which affects
the "color & fonts"

To help us make this distinciton, all "class" attributes are replaced
with "id" in index.html